refactor(cli): decompose the API client
Move transport, auth, graph, work-item, memory, source, and admin methods
into direct implementation owners behind the stable SibylClient surface.

Preserve all public signatures, request bodies, retry semantics, pending
writes, idempotency keys, and structured error envelopes.
